Add space values
Our build script is already configured to generate Sass variables and
CSS custom properties from our tokens, so adding our space scale is a
matter of adding a new properties file.

It's worth noting that we're following Style Dictionary's CTI (Category
/ Type / Item) structure [1] here, so using "size" as a top-level key
("category") has a special meaning: when converting to Scss it will be
treated as being in `rem` units by default (which is what we want).

[1] https://amzn.github.io/style-dictionary/#/properties?id=category-type-item