Add motion values (for animations / transitions)
Let's specify some values for durations and easing functions. These are
things that are currently specified separately in both JS (in
`utils/transitions`) and CSS, so this is somewhere where platform-
agnostic design tokens are particularly useful, since they provide a
single source of truth for both JS and CSS.

It's worth noting that because we're following Style Dictionary's CTI
(Category / Type / Item) structure [1], "time" has a special meaning
here: it provides Style Dictionary with a hint that these numbers are
times in milliseconds. When output as CSS, Style Dictionary specifies
these in seconds (which is fine), but in JS we'll probably want to
retain the raw millisecond values. We'll do this in the next commit.

[1] https://amzn.github.io/style-dictionary/#/properties?id=category-type-item