Create new JS `module.exports` format to allow direct import elsewhere
In the previous commit, we generated JS files with values specified
via `export const`

In order to install this as a package in node_modules, we'll need to use
the `module.exports` format for our JavaScript constants instead.

This is because the ES6 `export const` format needs to be run through
Babel. But if we're using the package directly this won't happen -
instead we need to format the JS ourselves.

This commit does that, via a custom formatter.

The contents of `duration.js`, `easing.js`, and `breakpoints.js` should
now be directly importable, via eg:

    import duration from 'fl-design-tokens/build/js/duration';

The values will then be available via, eg. `duration.short`

Co-authored-by: Tom Marshall <tom.marshall@futurelearn.com>