Add breakpoints as Sass map
In the FL app, we currently specify breakpoints as a map in Sass, and
then we have various helper functions that operate on this map.

The easiest way of tokenising these values is to allow the tokens to
generate a map in the exact format that our existing code expects.

To do this, we need to drop the "category" part of the name, which will
require a custom transform. This transform (`name/ti/kebab`) is a very
slightly modified version of the existing `name/cti/kebab` transform [1]

We'll be able to use this Transform to generate similar maps for,
eg. colours

[1] https://github.com/amzn/style-dictionary/blob/42fedce772dcf7b4beff355a9719f5ffaa378fe3/lib/common/transforms.js#L187