Add Sass map for spacings
Let's add the "size/rem" transform so spacings (being in the "size"
category) are output in `rem` units when rendered as a Sass map.

The keys here are slightly different from the keys in the FL app: there
we just use numbers, here we have, eg. `space-1`. Rather than wrangle
with the formatting, it should be easy enough to address this just by
tweaking the `space` function that uses this map.