Remove "fl" prefix from colour names
This was introduced in order to maintain the same separation between the
"brand" and "off-brand" palettes that currently exists in the main app.

However, @urbanautomaton suggested [1] that for the brand palette it
would be nice to avoid the prefix, and I'm inclined to agree.

I've removed the prefix by simply moving the colours up a level in their
properties file. This does mean that this file diverges from Style
Dictionary's recommended "CTI" (Category / Type / Item) structure, but
in practice the only difference this seems to make is that the
`name/human` transform that we're using to create human-friendly names
for the colours when generating a Sketch palette no longer works.

This is relatively easy to fix: I've made a custom transform that
formats the names appropriately (replacing '-' with ' ' and upper-casing
the first letter of each word).

[1] https://github.com/futurelearn/design-tokens/commit/b5db89116179f5b5436638a22f8144c516e53399#r37095115