Use US spelling of "color" throughout the Design Tokens repo
At some point in the distant past, a decision was made to use the UK
spelling of "colour" in our app. However, given that "color" is a
keyword in both CSS and Style Dictionary, using UK spelling in some
places when we're obliged to use US spelling in others is a source of
inconsistency and confusion.

Rather than allow this confusion to spread into our design tokens, let's
just use the US / CSS spelling consistently throughout the whole repo.