Add "ts/enums" format
We've started using TypeScript in the main FL app. One thing that TS
provides is an Enum data type, which can be used to provide a list of
allowable values.

This sounds very useful for places where we're using design tokens:
appropriate Enums would allow us to check that the values we're using
are indeed from our design tokens.

So let's introduce a new "ts/enums" format that will output TypeScript
enums for any property that it is provided with. The resulting output
looks something like:

```
export enum Breakpoint {
  TypeScaleMedium = 680,
  TypeScaleLarge = 1695,
  Small = 480,
  Medium = 648,
  Large = 900,
  Xlarge = 1280,
  Xxlarge = 1600,
  Comments = 1392,
};
```