Update Sketch palette tokens
Following the upgrade to Style Dictionary v3.0.0, the colour values here
have a higher precision.

This `.sketchpalette` file is a syntax that can be imported into Sketch
as a colour palette. I'm presuming the increased precision here reflects
some change in the rounding that happens in the conversion from the hex
values that we use to specify our colours and the RGB format that
Sketch uses.