Rename "properties" to "tokens"
Prior to v3, Style Dictionary used the term "properties" to describe the
values that it operates on. From v3.0.0 onwards, they are calling these
"tokens" instead, for consistency with common usage / understanding.

While the "property"-related syntax still works for backwards-
compatibility reasons, let's update our code to reflect the new
terminology.

This involved renaming the "properties" folder to "tokens", updating
the places in `build.js` that called `allProperties` to call `allTokens`
(which is just an alias), and renaming a few variables from, eg. `prop`
to `token`